requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
aiohttp
mysql-connector-python
psutil
beautifulsoup4
pyyaml